Overhaul syslog implementation
This change brings a distinction between the management of syslog in this
puppet repo and and the management of syslog by the OS.  The basic
philosophy followed is:

- Do not touch the OS-provided syslog files or the management of them
  (including logrotate).
- Provide the minimum amount of configuration required to set up our
  additional logging.
- Rotate the additional log files explicitly with logrotate, but do not
  restart syslog - trust the OS to do this appropriately.

Specifics:
- All syslog files are kept in /var/log/sysmgt.
- A separate file is kept for each syslog facility, with an additional file
  containing all facilities (/var/log/sysmgt/all)

The aims of this setup are:
- Provide granular logging for use by scripts in this puppet repository.
- Ensure smooth upgrades of OS-provided syslog and logrotate configuration.
- Ensure the system remains familiar to support personnel who expect to
  find the OS's default logging configuration in place.

This setup may not be appropriate in environments with high syslog volumes
(e.g. ISP mail servers), given that syslog will potentially log each
message in up to 3 places.  Please be wise about where you implement it!